SagaAlta_funciona
Había un error con el tipo de valor de idPelicula, tenia que estar en tinynt unsigned y estaba en medium. Y después quedaba en valor 0, lo que en el HttpPost de de SagaController le había que agregarle vmsaga.Saga.IdPelicula = vmsaga.IdPelicula;. Para que llame al idPelicula con un valor.

+++ SQL/procedure.sql
@@ -212,10 +212,10 @@ end $$
 
 DELIMITER $$
 /*procedure de saga */
- drop procedure if exists insSaga $$
+ drop procedure if exists insSaga $$ 
  Create procedure insSaga(out   unidsaga TINYINT UNSIGNED,
                                 unNumero_Saga TINYINT UNSIGNED,
-                                unidpelicula MEDIUMINT unsigned,
+                                unidpelicula TINYINT unsigned,
                                 unnombre varchar (50) )
 begin
  insert into Saga(Numero_Saga,idpelicula,nombre)
@@ -234,7 +234,7 @@ DELIMITER $$
 drop PROCEDURE if EXISTS UpdSaga $$
 CREATE PROCEDURE UpdSaga(out unidsaga TINYINT UNSIGNED,
                             unNumero_Saga TINYINT UNSIGNED,
-                            unidpelicula MEDIUMINT unsigned,
+                            unidpelicula TINYINT unsigned,
                             unnombre varchar(50))
 begin 
         update Saga
